pyproject.toml
[project]
name = "bot-telegram-gastos"
version = "0.1.0"
description = "Bot de Telegram para manejo de finanzas personales y compartidas"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.6",
]
